langfuse-setup/langgraph-agent/backend/requirements.txt
fastapi==0.115.5
uvicorn==0.34.0
langgraph==0.2.59
langchain-core==0.3.28
langchain-openai==0.2.14
langfuse==3.11.0
pydantic==2.10.3
python-dotenv==1.0.1
httpx==0.27.0
